Create Graph Class
Graph class to implement the graph. Add_edges function adds the edges for a directed graph and labels the edge, based on the vertices it connects, with its corresponding sound. The sound is inputted by the user in the add_edge function. All of the other functions remained the same.