Add files via upload
Working Submission for Part A. The prior NoneType error was not due to improper return on recursion; it was a problem in improperly updating the pathTaken list. Fixed. See comments in code for a more detailed description of how it works. This should be the copy to use for your own work, not the AF file. I'll leave that in so Professor, may see the progress.