Add files via upload
The randomTestMethod basically comes up with a random directed graph with a maximum of 30 vertices. This can be changed to be fewer or greater in the function. It then selects vertices at random and assigns them a random sequence of 'sounds.' The current options for this are 'a,b,c,d' but again this can be changed to include more or fewer 'sounds' however the more sounds there are the less likely there is for the sound sequence being searched to exist.